move Recon generated code to org.apache.ozone.recon.schema.generated
