+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/persistence/TestReconWithDifferentSqlDBs.java
@@ -19,7 +19,7 @@ package org.apache.hadoop.ozone.recon.persistence;
 
 import static org.apache.hadoop.ozone.recon.ReconControllerModule.ReconDaoBindingModule.RECON_DAO_LIST;
 import static org.apache.ozone.recon.codegen.SqlDbUtils.SQLITE_DRIVER_CLASS;
-import static org.apache.ozone.recon.schema.Tables.RECON_TASK_STATUS;
+import static org.apache.ozone.recon.schema.generated.Tables.RECON_TASK_STATUS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -30,8 +30,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.SQLException;
 import java.util.stream.Stream;
-import org.apache.ozone.recon.schema.tables.daos.ReconTaskStatusDao;
-import org.apache.ozone.recon.schema.tables.pojos.ReconTaskStatus;
+import org.apache.ozone.recon.schema.generated.tables.daos.ReconTaskStatusDao;
+import org.apache.ozone.recon.schema.generated.tables.pojos.ReconTaskStatus;
 import org.jooq.SQLDialect;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
